OPEN-00: Add Swift crypto module and update workflows (#74)
## Changes
- Added Swift crypto module for unified binding release.
- Updated workflows to support the new changes.

## Breaking Changes
- None

+++ core-modules-kotlin/healthcard-testkit/src/jvmMain/kotlin/de/gematik/openhealth/healthcard/replay/ReplayHarnessJvm.kt
@@ -67,7 +67,7 @@ actual class SecureChannelHandle internal constructor(
         return VerifyPinOutcome.valueOf(result.outcome.name)
     }
 
-    actual fun getRandom(length: UInt): ByteArray = secureChannel.getRandom(length)
+    actual fun getRandom(length: UInt): ByteArray = secureChannel.getRandom(length.toInt())
 
     actual fun readVsd(): ByteArray = secureChannel.readVsd()
